DeepLearning/requirements.txt
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/nix/avx2
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/nix/generic
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/generic
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
joblib==0.17.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.3
Pillow==7.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.11.0
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.11.2
